refactor(core): major refactor of the code, powerctl completely transfered into power manager structure [no changelog]
