+++ core/embed/sys/power_manager/fuel_gauge/battery_data_jyhpfl333838.h
