+++ core/embed/sys/power_manager/fuel_gauge/battery_model.c
