+++ core/embed/sys/power_manager/fuel_gauge/fuel_gauge.c
