+++ core/embed/sys/power_manager/npm1300/npm1300_defs.h
