+++ core/embed/sys/power_manager/stm32u5/power_monitoring.c
@@ -22,9 +22,9 @@
 #include <sys/systick.h>
 #include <trezor_rtl.h>
 
-#include "../../powerctl/fuel_gauge/fuel_gauge.h"
-#include "../../powerctl/npm1300/npm1300.h"
-#include "../../powerctl/stwlc38/stwlc38.h"
+#include "../fuel_gauge/fuel_gauge.h"
+#include "../npm1300/npm1300.h"
+#include "../stwlc38/stwlc38.h"
 #include "power_manager_internal.h"
 
 void pm_monitor_power_sources(void) {
