+++ core/embed/sys/power_manager/stm32u5/power_states.c
@@ -18,13 +18,17 @@
  */
 
 #include <io/backlight.h>
+#ifdef USE_BUTTON
 #include <io/button.h>
+#endif
+#ifdef USE_RGB_LED
 #include <io/rgb_led.h>
+#endif
 #include <sys/bootutils.h>
 #include <sys/systick.h>
 #include <sys/systimer.h>
 
-#include "../../powerctl/npm1300/npm1300.h"
+#include "../npm1300/npm1300.h"
 #include "power_manager_internal.h"
 
 // State handler lookup table
@@ -224,11 +228,13 @@ pm_internal_state_t pm_handle_state_startup_rejected(pm_driver_t* drv) {
 }
 
 void pm_enter_charging(pm_driver_t* drv) {
+
+#ifdef USE_RGB_LED
   // Initialize RGB
   rgb_led_init();
-
-  // Set RGB LED to blue
   rgb_led_set_color(0x0000FF);
+#endif
+
 }
 
 pm_internal_state_t pm_handle_state_charging(pm_driver_t* drv) {
@@ -255,8 +261,12 @@ pm_internal_state_t pm_handle_state_charging(pm_driver_t* drv) {
 }
 
 void pm_exit_charging(pm_driver_t* drv) {
+
+#ifdef USE_RGB_LED
   // Turn off RGB LED
   rgb_led_set_color(0x0);
+#endif
+
 }
 
 pm_internal_state_t pm_handle_state_hibernate(pm_driver_t* drv) {
