+++ core/embed/sys/power_manager/stwlc38/nvm_data_cfg_v4_cut_1_3.h
