+++ core/embed/sys/power_manager/stwlc38/nvm_data_fw_patch_v4.h
