+++ core/embed/sys/power_manager/stwlc38/stwlc38_internal.h
