+++ core/site_scons/models/T3W1/trezor_t3w1_revA.py
@@ -234,15 +234,18 @@ def configure(
     ]
 
     sources += [
-        "embed/sys/powerctl/npm1300/npm1300.c",
-        "embed/sys/powerctl/stwlc38/stwlc38.c",
-        "embed/sys/powerctl/stwlc38/stwlc38_patching.c",
-        "embed/sys/powerctl/stm32u5/powerctl.c",
-        "embed/sys/powerctl/stm32u5/powerctl_suspend.c",
-        "embed/sys/powerctl/wakeup_flags.c",
+        "embed/sys/power_manager/stm32u5/power_manager.c",
+        "embed/sys/power_manager/stm32u5/power_monitoring.c",
+        "embed/sys/power_manager/stm32u5/power_states.c",
+        "embed/sys/power_manager/stm32u5/power_control.c",
+        "embed/sys/power_manager/npm1300/npm1300.c",
+        "embed/sys/power_manager/fuel_gauge/fuel_gauge.c",
+        "embed/sys/power_manager/fuel_gauge/battery_model.c",
+        "embed/sys/power_manager/stwlc38/stwlc38.c",
+        "embed/sys/power_manager/stwlc38/stwlc38_patching.c",
     ]
-    paths += ["embed/sys/powerctl/inc"]
-    defines += [("USE_POWERCTL", "1")]
+    paths += ["embed/sys/power_manager/inc"]
+    defines += [("USE_POWER_MANAGER", "1")]
 
     env.get("ENV")["LINKER_SCRIPT"] = linker_script
